Restructure project for worker and stegtool
Add multiple build targets with a unified code base.

Plan is to have a server, worker, and stegtool binary.
The server handles requests, creates jobs, and pushes them to the db, and to rabbit.
The worker listens for jobs on rabbit, then calls the stegtool to perform the steg.
The worker is separate from stegtool so the image processing can be run inside nsjail, as I don't trust user input.